refactor mirgecompare.py into library + script
